requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
Pillow>=9.0.0
opencv-python>=4.8.0
decord>=0.6.0
accelerate>=0.20.3
einops>=0.6.1
timm>=0.9.2
python-dotenv>=1.0.0
pytubefix>=3.7.12
youtube-transcript-api>=0.6.1
predictionguard>=0.1.0
tqdm>=4.65.0
datasets>=2.12.0
langchain-core>=0.1.0
lancedb>=0.3.0
mm-rag>=0.1.0
bridgetower-embeddings>=0.1.0
anthropic 